// Grupo 2/Fuentes/Cliente/src/app/interfaces/listaSiniestroI.ts
 export interface ListaSiniestroI{
    hechoGenerador: string;
    nombreCobertura: string;
    nombrePoliza: string;
    vehiculoMarca: string;
    vehiculoAnio: number;
    vehiculoModelo: string;
    vehiculoVersion: string;
    vehiculoPatente: string;
    fecha: string;
    hora: string;
    descripcion: string;
    idSiniestro: number;
    expanded: boolean;
    modificable: boolean;
    localidad: string 
 }
    